+++ input.go
@@ -1,0 +1,54 @@
+package kairos
+
+import (
+	"context"
+	"fmt"
+	"net"
+	"net/http"
+
+	"golang.org/x/xerrors"
+)
+
+func (k *kairosRestClient) GetInputs(ctx context.Context) ([]Input, error) {
+	// エンドポイントの設定
+	ep := fmt.Sprintf("http://%s/inputs", net.JoinHostPort(k.ip, k.port))
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, ep, nil)
+	if err != nil {
+		return nil, xerrors.Errorf("Failed to create request: %w", err)
+	}
+
+	var payload []Input
+	if err := k.doRequest(req, &payload); err != nil {
+		return nil, xerrors.Errorf("Failed to do request: %w", err)
+	}
+
+	return payload, nil
+}
+
+type InputIdentifier interface {
+	~int | ~string
+}
+
+func getInput[T InputIdentifier](ctx context.Context, k *kairosRestClient, input T) (*Input, error) {
+	// エンドポイントの設定
+	ep := fmt.Sprintf("http://%s/inputs/%v", net.JoinHostPort(k.ip, k.port), input)
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, ep, nil)
+	if err != nil {
+		return nil, xerrors.Errorf("Failed to create request: %w", err)
+	}
+
+	var payload Input
+	if err := k.doRequest(req, &payload); err != nil {
+		return nil, xerrors.Errorf("Failed to do request: %w", err)
+	}
+
+	return &payload, nil
+}
+
+func (k *kairosRestClient) GetInputByID(ctx context.Context, id string) (*Input, error) {
+	return getInput(ctx, k, id)
+}
+
+func (k *kairosRestClient) GetInputByNumber(ctx context.Context, number int) (*Input, error) {
+	return getInput(ctx, k, number)
+}
