+++ snapshot.go
@@ -1,0 +1,7 @@
+package kairos
+
+import "context"
+
+func (k *kairosRestClient) PatchSnapshot(ctx context.Context) error {
+	panic("TODO")
+}
